renaming back and apply changes
