+++ README.md
@@ -44,7 +44,7 @@ hostname: bar
 Add this line to your application's Gemfile:
 
 ```ruby
-gem 'sidekiq-alive-next'
+gem 'sidekiq_alive'
 ```
 
 And then execute:
@@ -53,7 +53,7 @@ And then execute:
 
 Or install it yourself as:
 
-    $ gem install sidekiq-alive-next
+    $ gem install sidekiq_alive
 
 ## Usage
 
@@ -282,7 +282,7 @@ Here is an example [rails app](https://github.com/arturictus/sidekiq_alive_examp
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at https://github.com/andrcuns/sidekiq-alive. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
+Bug reports and pull requests are welcome on GitHub at https://github.com/arturictus/sidekiq_alive. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
 
 ## License
